# docs/swagger.yaml
openapi: 3.0.2
info:
  title: Games-API
  version: 1.0.0
  description: '### This is a Swagger UI for our Games API'
paths:
  /games:
    get:
      requestBody: {}
      tags:
        - games
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              examples:
                List of games:
                  value:
                    - Witcher 3
                    - Cyberpunk 2077
          description: ':#'
      summary: Get a list for all games
      description: Returns a list of all games in API
    post:
      requestBody:
        description: A new game to be created
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/error'
            examples:
              error:
                value:
                  error: error
        required: true
      tags:
        - new-game
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/new-game'
          description: Game added successfuly.
        '400':
          description: One or all params are missing.
      summary: Creates a game
      description: Creates a new game
  '/games/{id}':
    get:
      tags:
        - details
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              examples:
                Full details of a given game:
                  value:
                    id: 4
                    name: Counter-Strike 2
                    price: 0
          description: Returns full details of a given id game
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Game /w+ was not found:
                  value: Game \\w+ was not found in game array.
          description: Game was not found in games array
      operationId: id
      summary: Get full details of a game
      description: Returns full details of a given id game
    delete:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/delete-game'
      tags:
        - delete-game
      parameters:
        -
          name: id
          schema:
            type: string
          in: path
          required: true
      responses:
        '204':
          description: 204 no content
        '404':
          description: 404 not found
      summary: Delete game
      description: Delete game by id
    parameters:
      -
        name: id
        schema:
          type: number
        in: path
        required: true
components:
  schemas:
    error:
      title: Root Type for error
      description: ''
      type: object
      properties:
        error:
          type: string
      example:
        error: Error message.
    new-game:
      title: Root Type for new-game
      description: ''
      required:
        - price
        - name
      type: object
      properties:
        name:
          type: string
        price:
          format: float
          type: number
      example:
        name: Tetris
        price: 0.99
    delete-game:
      description: ''
      required:
        - id
      type: object
      properties:
        id:
          description: ''
          type: string
      example:
        id: 1